Change in location of the config file and update in get-secrets.sh
1. Changes the location of the config file from
tmp/config.json to `conf/config.json`. The instructions related
to config.json and config.json.sample will be given in the
README file (see separate branch).

2. We no longer need to set our own developers CAPI keys in the config file.
The script get-secrets.sh has been updated to generate the config file from
the one in the `alexa-config` bucket. For this, correct JANUS credentials
need to be available locally as they are used to access the bucket.

To detect whether we are on local or being deployed on TeamCity, the code
tests the TEAMCITY_BRANCH environment variable.