pplying feross/standard (Part 1)
**Section 1**

This commit applies to the code base the standard defined here: http://standardjs.com/

The package.json file has been updated to declare a new dependency, but the test script has not been updated yet.

The test script is currently

```
"test": "gulp compile && tap test"
```

and it will become in a later commit

```
"test": "standard && gulp compile && tap test"
```

We are not yet including `standard` as part of the test because the test would fail for the moment. See reasons below

**Section 2: Places where the code need to be updated but has not changed yet**

1. Anything under the folder `test` has not changed. This will be done in a later commit as I did not feel like changing the code and the tests in the same commit.

2. Places where the script highlighted that an exception was not handled have not been modified. This will come in a later commit.

3. "src/capiQueryBuilder.js:109:41: Unnecessary use of conditional expression for default assignment." Will be done in a later commit.

**Section 3: Summary**

This commit is the "trivial" part of the migration.